make $.download(FOLDER) extendable
... mainly it's needed to override file download in `selenide-selenoid` project.
