+++ src/main/java/com/codeborne/selenide/DownloadsFolder.java
@@ -1,5 +1,6 @@
 package com.codeborne.selenide;
 
+import com.codeborne.selenide.files.DownloadedFile;
 import com.codeborne.selenide.files.FileFilter;
 
 import javax.annotation.CheckReturnValue;
@@ -17,6 +18,7 @@ import static java.util.Collections.emptyList;
 import static java.util.Collections.emptyMap;
 import static java.util.Locale.ROOT;
 import static java.util.stream.Collectors.joining;
+import static java.util.stream.Collectors.toList;
 import static java.util.stream.Collectors.toMap;
 import static org.apache.commons.io.FilenameUtils.getExtension;
 
@@ -39,6 +41,22 @@ public abstract class DownloadsFolder {
   public List<File> files() {
     File[] files = folder.listFiles();
     return files == null ? emptyList() : asList(files);
+  }
+
+  public List<DownloadedFile> filesNewerThan(long modifiedAfterTs) {
+    return files().stream()
+      .filter(File::isFile)
+      .filter(file -> isFileModifiedLaterThan(file, modifiedAfterTs))
+      .map(file -> new DownloadedFile(file, emptyMap()))
+      .collect(toList());
+  }
+
+  /**
+   * Depending on OS, file modification time can have seconds precision, not milliseconds.
+   * We have to ignore the difference in milliseconds.
+   */
+  static boolean isFileModifiedLaterThan(File file, long timestamp) {
+    return file.lastModified() - timestamp >= -1000L;
   }
 
   public abstract void cleanupBeforeDownload();
